+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for searching projects
 - Support for searching sections
+- Support for searching labels
 
 ## [3.1.0] - 2025-05-07